requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.7
langchain-community>=0.2.0
SQLAlchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0